Large re-organization of directory structure. Specialized containers are now suffixed with OfType, as in TreapSetOfInt (the cross-isolate friendly version of TreapSet<int>)
