+++ lib/deeply_immutable.dart
@@ -1,0 +1,16 @@
+// Copyright 2025, kasper@byolimit.com
+// SPDX-License-Identifier: BSD-3-Clause
+library;
+
+export 'src/deeply_immutable/implicit_treap_of_double.dart';
+export 'src/deeply_immutable/implicit_treap_of_int.dart';
+export 'src/deeply_immutable/implicit_treap_of_string.dart';
+export 'src/deeply_immutable/treap_list_of_double.dart';
+export 'src/deeply_immutable/treap_list_of_int.dart';
+export 'src/deeply_immutable/treap_list_of_string.dart';
+export 'src/deeply_immutable/treap_of_double.dart';
+export 'src/deeply_immutable/treap_of_int.dart';
+export 'src/deeply_immutable/treap_of_string.dart';
+export 'src/deeply_immutable/treap_set_of_double.dart';
+export 'src/deeply_immutable/treap_set_of_int.dart';
+export 'src/deeply_immutable/treap_set_of_string.dart';
