+++ lib/src/debug_node.dart
@@ -1,6 +1,5 @@
-// Copyright 2024 - 2024, kasper@byolimit.com
+// Copyright 2024 - 2025, kasper@byolimit.com
 // SPDX-License-Identifier: BSD-3-Clause
-
 import 'dart:math';
 
 import 'package:meta/meta.dart';
