+++ lib/treap.dart
@@ -1,8 +1,16 @@
-// Copyright 2021 - 2024, kasper@byolimit.com
+// Copyright 2021 - 2025, kasper@byolimit.com
 // SPDX-License-Identifier: BSD-3-Clause
 library;
 
+export 'src/implicit_treap.dart';
+export 'src/implicit_treap_base.dart';
+export 'src/node.dart';
+export 'src/debug_node.dart';
+export 'src/treap.dart';
 export 'src/treap_base.dart';
 export 'src/treap_list.dart';
+export 'src/treap_list_base.dart';
 export 'src/treap_map.dart';
+export 'src/treap_map_base.dart';
 export 'src/treap_set.dart';
+export 'src/treap_set_base.dart';
