+++ test/deeply_immutable_node_test.dart
@@ -1,8 +1,10 @@
-// Copyright 2024 - 2024, kasper@byolimit.com
+// Copyright 2024 - 2025, kasper@byolimit.com
 // SPDX-License-Identifier: BSD-3-Clause
 
 import 'package:test/test.dart';
-import 'package:treap/src/deeply_immutable_node.dart';
+import 'package:treap/src/deeply_immutable/double_node.dart';
+import 'package:treap/src/deeply_immutable/int_node.dart';
+import 'package:treap/src/deeply_immutable/string_node.dart';
 
 void main() {
   group('IntNode', () {
