+++ test/treap_test.dart
@@ -2,12 +2,11 @@
 // SPDX-License-Identifier: BSD-3-Clause
 import 'dart:math';
 
-import 'package:treap/src/deeply_immutable_node.dart';
+import 'package:test/test.dart';
+import 'package:treap/src/deeply_immutable/int_node.dart';
 import 'package:treap/src/hash.dart';
 import 'package:treap/src/immutable_node.dart';
-import 'package:treap/src/node.dart';
 import 'package:treap/treap.dart';
-import 'package:test/test.dart';
 
 import 'util.dart';
